backend/requirements.txt
django
psycopg2-binary
django-environ
djangorestframework